Removing comments etc.

+++ src/main/java/com/example/dota2/model/Hero.java
@@ -1,7 +1,6 @@
 package com.example.dota2.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 import java.util.List;
 
@@ -16,48 +15,48 @@ public class Hero {
     private String attackType;
     @JsonProperty("1_pick")
     //Herald = 1
-    private int pick1;
+    private double pick1;
     @JsonProperty("2_pick")
     //Guardian = 2
-    private int pick2;
+    private double pick2;
     @JsonProperty("3_pick")
     //Crusader = 3
-    private int pick3;
+    private double pick3;
     @JsonProperty("4_pick")
     //Archon = 4
-    private int pick4;
+    private double pick4;
     @JsonProperty("5_pick")
     //Legend = 5
-    private int pick5;
+    private double pick5;
     @JsonProperty("6_pick")
     //Ancient = 6
-    private int pick6;
+    private double pick6;
     @JsonProperty("7_pick")
     //Divine = 7
-    private int pick7;
+    private double pick7;
     @JsonProperty("8_pick")
     //Immortal = 8
-    private int pick8;
+    private double pick8;
     @JsonProperty("pro_pick")
-    private int proPicked;
+    private double proPicked;
     @JsonProperty("pro_win")
-    private int proWins;
+    private double proWins;
     @JsonProperty("1_win")
-    private int win1;
+    private double win1;
     @JsonProperty("2_win")
-    private int win2;
+    private double win2;
     @JsonProperty("3_win")
-    private int win3;
+    private double win3;
     @JsonProperty("4_win")
-    private int win4;
+    private double win4;
     @JsonProperty("5_win")
-    private int win5;
+    private double win5;
     @JsonProperty("6_win")
-    private int win6;
+    private double win6;
     @JsonProperty("7_win")
-    private int win7;
+    private double win7;
     @JsonProperty("8_win")
-    private int win8;
+    private double win8;
     @JsonProperty("heroComplexity")
     private int heroComplexity;
     @JsonProperty("heroPosition")
@@ -108,147 +107,147 @@ public class Hero {
         this.attackType = attackType;
     }
 
-    public int getPick1() {
+    public double getPick1() {
         return pick1;
     }
 
-    public void setPick1(int pick1) {
+    public void setPick1(double pick1) {
         this.pick1 = pick1;
     }
 
-    public int getPick2() {
+    public double getPick2() {
         return pick2;
     }
 
-    public void setPick2(int pick2) {
+    public void setPick2(double pick2) {
         this.pick2 = pick2;
     }
 
-    public int getPick3() {
+    public double getPick3() {
         return pick3;
     }
 
-    public void setPick3(int pick3) {
+    public void setPick3(double pick3) {
         this.pick3 = pick3;
     }
 
-    public int getPick4() {
+    public double getPick4() {
         return pick4;
     }
 
-    public void setPick4(int pick4) {
+    public void setPick4(double pick4) {
         this.pick4 = pick4;
     }
 
-    public int getPick5() {
+    public double getPick5() {
         return pick5;
     }
 
-    public void setPick5(int pick5) {
+    public void setPick5(double pick5) {
         this.pick5 = pick5;
     }
 
-    public int getPick6() {
+    public double getPick6() {
         return pick6;
     }
 
-    public void setPick6(int pick6) {
+    public void setPick6(double pick6) {
         this.pick6 = pick6;
     }
 
-    public int getPick7() {
+    public double getPick7() {
         return pick7;
     }
 
-    public void setPick7(int pick7) {
+    public void setPick7(double pick7) {
         this.pick7 = pick7;
     }
 
-    public int getPick8() {
+    public double getPick8() {
         return pick8;
     }
 
-    public void setPick8(int pick8) {
+    public void setPick8(double pick8) {
         this.pick8 = pick8;
     }
 
-    public int getProPicked() {
+    public double getProPicked() {
         return proPicked;
     }
 
-    public void setProPicked(int proPicked) {
+    public void setProPicked(double proPicked) {
         this.proPicked = proPicked;
     }
 
-    public int getProWins() {
+    public double getProWins() {
         return proWins;
     }
 
-    public void setProWins(int proWins) {
+    public void setProWins(double proWins) {
         this.proWins = proWins;
     }
 
-    public int getWin1() {
+    public double getWin1() {
         return win1;
     }
 
-    public void setWin1(int win1) {
+    public void setWin1(double win1) {
         this.win1 = win1;
     }
 
-    public int getWin2() {
+    public double getWin2() {
         return win2;
     }
 
-    public void setWin2(int win2) {
+    public void setWin2(double win2) {
         this.win2 = win2;
     }
 
-    public int getWin3() {
+    public double getWin3() {
         return win3;
     }
 
-    public void setWin3(int win3) {
+    public void setWin3(double win3) {
         this.win3 = win3;
     }
 
-    public int getWin4() {
+    public double getWin4() {
         return win4;
     }
 
-    public void setWin4(int win4) {
+    public void setWin4(double win4) {
         this.win4 = win4;
     }
 
-    public int getWin5() {
+    public double getWin5() {
         return win5;
     }
 
-    public void setWin5(int win5) {
+    public void setWin5(double win5) {
         this.win5 = win5;
     }
 
-    public int getWin6() {
+    public double getWin6() {
         return win6;
     }
 
-    public void setWin6(int win6) {
+    public void setWin6(double win6) {
         this.win6 = win6;
     }
 
-    public int getWin7() {
+    public double getWin7() {
         return win7;
     }
 
-    public void setWin7(int win7) {
+    public void setWin7(double win7) {
         this.win7 = win7;
     }
 
-    public int getWin8() {
+    public double getWin8() {
         return win8;
     }
 
-    public void setWin8(int win8) {
+    public void setWin8(double win8) {
         this.win8 = win8;
     }
